:art: Separate `cib` and `cib_nexus` libraries
Problem:
- `cib_nexus` is a separate sub-library from `cib`, but the headers are
  intermingled in the same directory.

Solution:
- Separate out the `cib_nexus` headers into the `nexus` include directory.
