+++ include/flow/dsl/seq.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include <cib/detail/runtime_conditional.hpp>
 #include <flow/dsl/subgraph_identity.hpp>
 #include <flow/dsl/walk.hpp>
+#include <nexus/detail/runtime_conditional.hpp>
 
 #include <stdx/tuple_algorithms.hpp>
