+++ include/flow/step.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
-#include <cib/detail/runtime_conditional.hpp>
-#include <cib/func_decl.hpp>
 #include <flow/dsl/subgraph_identity.hpp>
 #include <flow/log.hpp>
+#include <nexus/detail/runtime_conditional.hpp>
+#include <nexus/func_decl.hpp>
 
 #include <stdx/compiler.hpp>
 #include <stdx/ct_string.hpp>
