+++ include/nexus/built.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <cib/builder_meta.hpp>
+#include <nexus/builder_meta.hpp>
 
 namespace cib {
 template <builder_meta ServiceMeta>
