+++ include/nexus/callback.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <cib/builder_meta.hpp>
+#include <nexus/builder_meta.hpp>
 
 #include <stdx/compiler.hpp>
 #include <stdx/panic.hpp>
