+++ include/nexus/detail/exports.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <cib/detail/config_item.hpp>
-#include <cib/detail/extend.hpp>
+#include <nexus/detail/config_item.hpp>
+#include <nexus/detail/extend.hpp>
 
 #include <stdx/tuple.hpp>
