+++ test/cib/builder_meta.cpp
@@ -1,4 +1,4 @@
-#include <cib/builder_meta.hpp>
+#include <nexus/builder_meta.hpp>
 
 #include <stdx/compiler.hpp>
