+++ test/cib/callback.cpp
@@ -1,4 +1,4 @@
-#include <cib/callback.hpp>
+#include <nexus/callback.hpp>
 
 #include <stdx/tuple.hpp>
