+++ test/cib/callback_uninit.cpp
@@ -1,5 +1,5 @@
-#include <cib/built.hpp>
-#include <cib/callback.hpp>
+#include <nexus/built.hpp>
+#include <nexus/callback.hpp>
 
 #include <stdx/panic.hpp>
