+++ test/flow/flow_cib_func.cpp
@@ -1,4 +1,4 @@
-#include <cib/func_decl.hpp>
+#include <nexus/func_decl.hpp>
 
 #include <string>
