+++ test/msg/handler_uninit.cpp
@@ -1,6 +1,6 @@
-#include <cib/built.hpp>
 #include <msg/message.hpp>
 #include <msg/service.hpp>
+#include <nexus/built.hpp>
 
 #include <stdx/panic.hpp>
