+++ test/msg/indexed_handler_uninit.cpp
@@ -1,6 +1,6 @@
-#include <cib/built.hpp>
 #include <msg/indexed_service.hpp>
 #include <msg/message.hpp>
+#include <nexus/built.hpp>
 
 #include <stdx/panic.hpp>
